Moved the rest of reviewtools.py into example_post_processing.py
This ensures that example-checker-ipynb uses the most recent code from epp.
The reviewtools.py file has been removed.
sys.path variables updated in example-checker.ipynb to grab correct code.
